==> src/6-add-elev-to-digitized.ipynb <==
## Add elevation data to digitized buildings

import os

from multiprocessing import Pool
from scipy import stats

import geopandas as gpd
import pandas as pd
import numpy as np

from shapely.geometry import MultiPolygon

def mode(x):
    mode, count = stats.mode(np.round(x))
    return mode[0]

base_path = "../data/Digitization/"
elev_path = "../data/lidar/las_to_vector/"

Read all data

sample_names = [x.rstrip('_PointHeight_cor.gpkg') for x in os.listdir(elev_path) if x.endswith('_PointHeight_cor.gpkg')]
samples = {}
for sample in sample_names:
    print(f'Reading sample {sample}')
    digi_name = sample.replace('_', '') + '_Digitiz.shp'
    elev_name = sample + '_PointHeight_cor.gpkg'
    samples[sample] = {
        'digi': gpd.read_file(os.path.join(base_path, digi_name)),
        'elev': gpd.read_file(os.path.join(elev_path, elev_name))
    }

Normalize column names

for fid, dfs in samples.items():
    cols_to_rename = {
        'Id': 'OBJECTID',
        'Area': 'SHAPE_Area'
    }
    dfs['digi'].rename(index=str, columns=cols_to_rename, inplace=True)
    dfs['digi'] = dfs['digi'][['OBJECTID', 'SHAPE_Area', 'Roof', 'geometry']]

## Assign points to buildings

Then, aggregate data by building and join back to digitized DFs to save

for k, v in samples.items():
    print(f'Working on {k}')
    v['elev'] = v['elev'].to_crs(v['digi'].crs)
    df = gpd.sjoin(v['elev'], v['digi'], op='within')
    agg_df = df.groupby('index_right').agg(
        {'z': [min, max, np.mean, np.median, mode]}
    )
    agg_df.columns = agg_df.columns.get_level_values(1)
    
    # Check if multipolygons are present
    # and cast all simple ones to multi if necessary
    geom_types = v['digi'].geometry.geom_type.unique()
    if ('MultiPolygon' in geom_types) and (len(geom_types) > 1):
        print('Polygon and Multipolygons detected. Casting all to Multi...')
        v['digi'].geometry = v['digi'].geometry.apply(lambda x: MultiPolygon([x]))
    
    v['digi'].join(agg_df).to_file(
        os.path.join(os.path.join(base_path, 'output'), f'{k}_digi_elev.gpkg'),
        driver='GPKG'
    )

## Quickfix: Remove points from vector files where height is Out Of Bounds

elev_path = "../data/lidar/las_to_vector/"
def process_save(file):
    print(file)
    fp = os.path.join(elev_path, file)
    gdf = gpd.read_file(fp)
    gdf.loc[(gdf['z'] >= 0) & (gdf['z'] <= 25), :].to_file(fp, driver='GPKG')

pool = Pool()
files = [x for x in os.listdir(elev_path) if x.endswith('.gpkg')]
pool.map(process_save, files)
